Move AddNoteForm
